requirements.txt
niquests
beautifulsoup4
fake-useragent
python-dotenv